RMSE accuracy tester + fixed recommend call
I fixed the recommend call so it would populate an array of recommendations for each user, giving an index for the movieID as the element of that array.

I also fixed the top 5 recommendations I believe, now it searches for the most common movieIDs and then prints them in descending order (from most recommended).

I am dubious about the validity of the accuracy tester, especially since we might have to rewrite it completely for the implementation of ALS without implicit.

Oh well, we'll worry about that later.